Remove rules/story
