+++ app/assets/javascripts/components/story/CollapsedStory/CollapsedStoryEstimate.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { estimateRule } from 'rules/story';
+
+const estimateRule = (estimate) => estimate > 0 ? estimate : '-';
 
 const CollapsedStoryEstimate = ({ estimate }) => (
   <span className='Story__estimated'>{estimateRule(estimate)}</span>
