+++ spec/javascripts/components/story/story_icon_spec.js
@@ -1,9 +1,9 @@
 import jasmineEnzyme from 'jasmine-enzyme';
 import React from 'react';
 import { shallow } from 'enzyme';
-import CollapsedStoryIcon from 'components/story/CollapsedStory/CollapsedStoryIcon';
+import StoryIcon from 'components/story/StoryIcon';
 
-describe('<CollapsedStoryIcon />', () => {
+describe('<StoryIcon />', () => {
   
   beforeEach(() => {
     jasmineEnzyme();
@@ -18,7 +18,7 @@ describe('<CollapsedStoryIcon />', () => {
   iconRules.forEach(({ storyType, icon, className }) => {
     describe(`when storyType = ${storyType}`,() => {
       it('renders the story icon', () => {
-        const wrapper = shallow(<CollapsedStoryIcon storyType={storyType} />);
+        const wrapper = shallow(<StoryIcon storyType={storyType} />);
 
         expect(wrapper.find('i')).toHaveClassName(`mi md-${className} md-16`);
         expect(wrapper.find('.Story__icon').text()).toContain(icon);
